+++ R/piece3d-rgl.R
@@ -36,6 +36,8 @@ piece3d <- function(piece_side = "tile_back", suit = NA, rank = NA, # nolint
                     alpha = 1.0, lit = FALSE,
                     shininess = 50.0, textype = NA) {
     assert_suggested("rgl")
+    if (is_angle(angle))
+        angle <- as.double(angle, "degrees")
 
     nn <- max(lengths(list(piece_side, suit, rank, x, y, z, angle, axis_x, axis_y, width, height, depth)))
     piece_side <- rep(piece_side, length.out = nn)